--- circuits/assets/circom/verification_key.json
@@ -1,94 +1,0 @@
-{
- "protocol": "groth16",
- "curve": "bn128",
- "nPublic": 1,
- "vk_alpha_1": [
-  "16428432848801857252194528405604668803277877773566238944394625302971855135431",
-  "16846502678714586896801519656441059708016666274385668027902869494772365009666",
-  "1"
- ],
- "vk_beta_2": [
-  [
-   "16348171800823588416173124589066524623406261996681292662100840445103873053252",
-   "3182164110458002340215786955198810119980427837186618912744689678939861918171"
-  ],
-  [
-   "19687132236965066906216944365591810874384658708175106803089633851114028275753",
-   "4920802715848186258981584729175884379674325733638798907835771393452862684714"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_gamma_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_delta_2": [
-  [
-   "10857046999023057135944570762232829481370756359578518086990519993285655852781",
-   "11559732032986387107991004021392285783925812861821192530917403151452391805634"
-  ],
-  [
-   "8495653923123431417604973247489272438418190587263600148770280649306958101930",
-   "4082367875863433681332203403145435568316851327593401208105741076214120093531"
-  ],
-  [
-   "1",
-   "0"
-  ]
- ],
- "vk_alphabeta_12": [
-  [
-   [
-    "5275725312362878540782176211860327475781113689246818544623830805017503247034",
-    "700769043921060225711174322502145319612473365595920873303028146383045646735"
-   ],
-   [
-    "16577533945604560505206253312979863148043263406037367789711279754781525822966",
-    "9408338099405950952721388539539775335199747835458172188116297223654842340186"
-   ],
-   [
-    "12663399896275491035004982800573482669934131767886952660443268164480899034271",
-    "4432711152773877173921024337047412943791122852326272337530740732443732395954"
-   ]
-  ],
-  [
-   [
-    "13121778684901402722679281862736806628725205381360313795132945954337708567513",
-    "9534744673358550231812045647241180985734073058548683258847806241019905135720"
-   ],
-   [
-    "21329152369227346659770815132468371951064045353268189088026893413117512652875",
-    "17209195434408943681049655974234541356066884378594227002358272904159790622854"
-   ],
-   [
-    "5346467096835895366917814311591075634165750361894629082277248282132405045579",
-    "15508364027636868967189209273443690126627947943852338696115233789046842639684"
-   ]
-  ]
- ],
- "IC": [
-  [
-   "2665582152853648751658299350874615127695381535641828682190197446149168953380",
-   "11621226065843605263824123089299496731848727836264368381166921078046445038357",
-   "1"
-  ],
-  [
-   "3677518104296008644609317043610726721245830832640288879392399632131240696730",
-   "3635398677697645421252340243263799546905321122555312519680240020101419885544",
-   "1"
-  ]
- ]
-}
